requirements.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
